+++ cddns/__main__.py
@@ -1,0 +1,3 @@
+from cddns.main import main
+
+main()
